requirements.txt
-i https://pypi.org/simple/
astroid==1.6.2
boto3==1.6.15
botocore==1.9.15
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
django-cors-headers==2.4.0
django==2.2.13
djangorestframework-simplejwt==3.3
djangorestframework==3.9.0
docutils==0.14
gunicorn==19.7.1
idna==2.8
isort==4.3.4
itypes==1.1.0
jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
markdown==3.0.1
markupsafe==1.1.0
mccabe==0.6.1
pillow==8.2.0
pygments==2.3.1
pyjwt==1.7.1
pylint==1.8.3
python-dateutil==2.6.1
python-decouple==3.1
pytz==2018.9
qrcode==6.0
requests==2.21.0
s3transfer==0.1.13
six==1.11.0
sqlparse==0.4.1
static3==0.7.0
uritemplate==3.0.0
urllib3==1.24.1
wrapt==1.10.11
django-rest-swagger